Settings WebUI: move all overrides to component-specific override module
Now that everything is defined in JS (with the move to es-modules), we can more declaratively define the overrides. This gives us an opportunity to make the overrides more readable, maintainable and expressed in the expected file name or the overridden component.
