+++ browser/resources/settings/brave_overrides/page_visibility.js
