+++ browser/resources/settings/brave_overrides/settings_menu.js
@@ -1,0 +1,226 @@
+// Copyright (c) 2020 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
+
+import {html} from 'chrome://resources/polymer/v3_0/polymer/polymer_bundled.min.js'
+import {RegisterStyleOverride, RegisterPolymerTemplateModifications} from 'chrome://brave-resources/polymer_overriding.js'
+import {loadTimeData} from '../i18n_setup.js'
+
+function createMenuElement (title, href, iconName) {
+  const menuEl = document.createElement('a')
+  menuEl.href = href
+  menuEl.innerHTML = `
+    <iron-icon icon="${iconName}"></iron-icon>
+    ${title}
+  `
+  return menuEl
+}
+
+function getMenuElement (templateContent, href) {
+  const menuEl = templateContent.querySelector(`a[href="${href}"]`)
+  if (!menuEl) {
+    console.error(`[Brave Settings Overrides] Could not find menu item '${href}'`)
+  }
+  return menuEl
+}
+
+RegisterStyleOverride(
+  'settings-menu',
+  html`
+    <style>
+      :host {
+        --brave-settings-menu-margin-v: 30px;
+        --brave-settings-menu-padding: 30px;
+        --settings-nav-item-color: #424242 !important;
+        position: sticky;
+        top: var(--brave-settings-menu-margin-v);
+        margin: 0 var(--brave-settings-menu-margin) !important;
+        max-height: calc(100vh - 56px - (var(--brave-settings-menu-margin-v) * 2) - (var(--brave-settings-menu-padding) * 2));
+        min-width: 172px;
+        border-radius: 6px;
+        background-color: #fff;
+        overflow-y: auto;
+        padding: 30px !important;
+      }
+
+      @media (prefers-color-scheme: dark) {
+        :host {
+          --settings-nav-item-color: #F4F4F4 !important;
+          border-color: transparent !important;
+          background-color: #161719;
+        }
+      }
+
+      a[href] {
+        font-weight: 400 !important;
+        margin: 0 20px 20px 0 !important;
+        margin-inline-start: 0 !important;
+        margin-inline-end: 0 !important;
+        padding-bottom: 0 !important;
+        padding-top: 0 !important;
+        padding-inline-start: 0 !important;
+      }
+
+      a[href].iron-selected {
+        color: var(--settings-nav-item-color) !important;
+        font-weight: 600 !important;
+      }
+
+      iron-icon {
+        color: var(--settings-nav-item-color);
+        margin-inline-end: 16px !important;
+        width: 20px;
+        height: 20px;
+      }
+
+      @media (prefers-color-scheme: dark) {
+        a[href].iron-selected iron-icon {
+          color: var(--settings-nav-item-color) !important;
+        }
+      }
+
+      a[href],
+      #advancedButton {
+        --cr-selectable-focus_-_outline: var(--brave-focus-outline) !important;
+      }
+
+      #advancedButton {
+        padding: 0 !important;
+        margin-top: 30px !important;
+        line-height: 1.25 !important;
+        border: none !important;
+      }
+
+      #advancedButton > iron-icon {
+        margin-inline-end: 0 !important;
+      }
+
+      #settingsHeader,
+      #advancedButton {
+        align-items: center !important;
+        font-weight: normal !important;
+        font-size: larger !important;
+        color: var(--settings-nav-item-color) !important;
+        margin-bottom: 20px !important;
+      }
+
+      #about-menu {
+        display: flex;
+        flex-direction: row;
+        align-items: flex-start;
+        justify-content: flex-start;
+        color: #c5c5d3 !important;
+        margin: 16px 0 0 0 !important;
+      }
+      .brave-about-graphic {
+        flex: 0;
+        flex-basis: 30%;
+        display: flex;
+        align-items: center;
+        justify-content: flex-start;
+        align-self: stretch;
+      }
+      .brave-about-meta {
+        flex: 1;
+      }
+      .brave-about-item {
+        display: block;
+      }
+    </style>
+  `
+)
+
+RegisterPolymerTemplateModifications({
+  'settings-menu': (templateContent) => {
+    // Add title
+    const titleEl = document.createElement('h1')
+    titleEl.id = 'settingsHeader'
+    titleEl.innerHTML = loadTimeData.getString('settings')
+    const topMenuEl = templateContent.querySelector('#topMenu')
+    if (!topMenuEl) {
+      console.error('[Brave Settings Overrides] Could not find topMenu element to add title after')
+    } else {
+      topMenuEl.insertAdjacentElement('afterbegin', titleEl)
+    }
+    // Advanced text
+    const advancedToggle = templateContent.querySelector('#advancedButton span')
+    if (!advancedToggle) {
+      console.error('[Brave Settings Overrides] Could not find advancedButton to modify text')
+    }
+    advancedToggle.innerText = loadTimeData.getString('braveAdditionalSettingsTitle')
+    // Add 'Get Started' item
+    const peopleEl = getMenuElement(templateContent, '/people')
+    const getStartedEl = createMenuElement(loadTimeData.getString('braveGetStartedTitle'), '/getStarted', 'brave_settings:get-started')
+    peopleEl.insertAdjacentElement('afterend', getStartedEl)
+    // Remove People item
+    peopleEl.remove()
+    // Move Appearance item
+    const appearanceBrowserEl = getMenuElement(templateContent, '/appearance')
+    getStartedEl.insertAdjacentElement('afterend', appearanceBrowserEl)
+    // Add New Tab item
+    const newTabEl = createMenuElement(loadTimeData.getString('braveNewTab'), '/newTab', 'brave_settings:new-tab')
+    appearanceBrowserEl.insertAdjacentElement('afterend', newTabEl)
+    // Add Sync and Help Tips item
+    const helpTipsEl = createMenuElement(loadTimeData.getString('braveHelpTips'), '/braveHelpTips', 'brave_settings:help')
+    if (loadTimeData.getBoolean('isSyncDisabled')) {
+      newTabEl.insertAdjacentElement('afterend', helpTipsEl)
+    } else {
+      const syncEl = createMenuElement(loadTimeData.getString('braveSync'), '/braveSync', 'brave_settings:sync')
+      newTabEl.insertAdjacentElement('afterend', syncEl)
+      syncEl.insertAdjacentElement('afterend', helpTipsEl)
+    }
+    // Add Shields item
+    const shieldsEl = createMenuElement(loadTimeData.getString('braveShieldsTitle'), '/shields',  'brave_settings:shields')
+    helpTipsEl.insertAdjacentElement('afterend', shieldsEl)
+    // Add Embed Blocking item
+    const embedEl = createMenuElement(loadTimeData.getString('socialBlocking'), '/socialBlocking', 'brave_settings:social-permissions')
+    shieldsEl.insertAdjacentElement('afterend', embedEl)
+    // Move search item
+    const searchEl = getMenuElement(templateContent, '/search')
+    embedEl.insertAdjacentElement('afterend', searchEl)
+    // Add Extensions item
+    const extensionEl = createMenuElement(loadTimeData.getString('braveDefaultExtensions'), '/extensions', 'brave_settings:extensions')
+    searchEl.insertAdjacentElement('afterend', extensionEl)
+    // Remove default Browser
+    const defaultBrowserEl = getMenuElement(templateContent, '/defaultBrowser')
+    defaultBrowserEl.remove()
+    // Remove Startup
+    const startupEl = getMenuElement(templateContent, '/onStartup')
+    startupEl.remove()
+    // Remove Accessibility :-(
+    const a11yEl = getMenuElement(templateContent, '/accessibility')
+    a11yEl.remove()
+    // Move autofill to advanced
+    const autofillEl = getMenuElement(templateContent, '/autofill')
+    const languagesEl = getMenuElement(templateContent, '/languages')
+    languagesEl.insertAdjacentElement('beforebegin', autofillEl)
+    // Move privacy to advanced
+    const privacyEl = getMenuElement(templateContent, '/privacy')
+    autofillEl.insertAdjacentElement('beforebegin', privacyEl)
+    // Move helptips to advanced
+    const printingEl = getMenuElement(templateContent, '/printing')
+    printingEl.insertAdjacentElement('afterend', helpTipsEl)
+    // Remove extensions link
+    const extensionsLinkEl = templateContent.querySelector('#extensionsLink')
+    if (!extensionsLinkEl) {
+      console.error('[Brave Settings Overrides] Could not find extensionsLinkEl to remove')
+    }
+    extensionsLinkEl.remove()
+    // Add version number to 'about' link
+    const aboutEl = templateContent.querySelector('#about-menu')
+    if (!aboutEl) {
+      console.error('[Brave Settings Overrides] Could not find about-menun element')
+    }
+    const aboutTitleContent = aboutEl.innerHTML
+    aboutEl.innerHTML = `
+      <div class="brave-about-graphic">
+        <iron-icon icon="brave_settings:full-color-brave-lion"><iron-icon>
+      </div>
+      <div class="brave-about-meta">
+        <span class="brave-about-item brave-about-menu-link-text">${aboutTitleContent}</span>
+        <span class="brave-about-item brave-about-menu-version">v ${loadTimeData.getString('braveProductVersion')}</span>
+      </div>
+    `
+  }
+})
