+++ browser/resources/settings/brave_overrides/settings_section.js
