--- browser/resources/settings/style_overrides/settings_basic_page.js
@@ -1,18 +1,0 @@
-// Copyright (c) 2020 The Brave Authors. All rights reserved.
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at http://mozilla.org/MPL/2.0/.
-
-import {html} from 'chrome://resources/polymer/v3_0/polymer/polymer_bundled.min.js'
-import {RegisterStyleOverride} from 'chrome://brave-resources/polymer_overriding.js'
-
-RegisterStyleOverride(
-  'settings-basic-page',
-  html`
-    <style>
-      :host {
-        min-width: 544px !important;
-      }
-    </style>
-  `
-)
